requirements.txt
pandas
python-dotenv
psycopg2-binary
google-api-python-client
google-auth-httplib2
google-auth-oauthlib